move 06.using-cards, add bot.recipe and README
