--- javascript_nodejs/06.using-cards/using-cards.bot
@@ -1,14 +1,0 @@
-{
-    "name": "using-cards",
-    "secretKey": "",
-    "services": [
-        {
-            "type": "endpoint",
-            "name": "using-cards",
-            "id": "http://localhost:3978/api/messages",
-            "appId": "",
-            "appPassword": "",
-            "endpoint": "http://localhost:3978/api/messages"
-        }
-    ]
-}
