+++ samples/javascript_nodejs/06.using-cards/.env
@@ -1,0 +1,2 @@
+botFilePath=using-cards.bot
+botFileSecret=
