+++ samples/javascript_nodejs/06.using-cards/README.md
@@ -1,0 +1,65 @@
+This sample shows how to create a bot that uses Rich Cards. This bot example uses [`restify`](https://www.npmjs.com/package/restify) and [`dotenv`](https://www.npmjs.com/package/dotenv). 
+
+# To try this sample
+- Clone the repository
+    ```bash
+    git clone https://github.com/microsoft/botbuilder-samples.git
+    ```
+- In a terminal, 
+    ```bash
+    cd samples\javascript_nodejs\06.using-cards
+    ```
+- [Optional] Update the .env file under samples\javascript_nodejs\06.using-cards with your botFileSecret
+    For Azure Bot Service bots, you can find the botFileSecret under application settings.
+- Install modules and start the bot
+    ```bash
+    npm i & npm start
+    ```
+    Alternatively you can also use nodemon via
+    ```bash
+    npm i & npm run watch
+    ```
+
+# Testing the bot using Bot Framework Emulator
+[Microsoft Bot Framework Emulator](https://github.com/microsoft/botframework-emulator) is a desktop application that allows bot developers to test and debug their bots on localhost or running remotely through a tunnel.
+
+- Install the Bot Framework Emulator from [here](https://aka.ms/botframework-emulator)
+
+## Connect to bot using Bot Framework Emulator V4
+- Launch Bot Framework Emulator
+- File -> Open Bot Configuration and navigate to samples\javascript_nodejs\06.using-cards folder
+- Select using-cards.bot file
+
+# Rich Cards
+A key to good bot design is to send interactive media, such as Rich Cards. There are several different types of Rich Cards, which are as follows:
+
+- Animation Card
+- Audio Card
+- Hero Card
+- Receipt Card
+- Signin Card
+- Thumbnail Card
+- Video Card
+
+When [designing the user experience][Design the user experience] developers should consider adding visual elements such as Rich Cards.
+
+# Deploy this bot to Azure
+You can use the [MSBot](https://github.com/microsoft/botbuilder-tools) Bot Builder CLI tool to clone and configure any services this sample depends on. 
+
+To install all Bot Builder tools - 
+```bash
+npm i -g msbot chatdown ludown qnamaker luis-apis botdispatch luisgen
+```
+
+To clone this bot, run
+```
+msbot clone services -f deploymentScripts/msbotClone -n <BOT-NAME> -l <Azure-location> --subscriptionId <Azure-subscription-id>
+```
+
+# Further reading
+- [Azure Bot Service Introduction](https://docs.microsoft.com/en-us/azure/bot-service/bot-service-overview-introduction?view=azure-bot-service-4.0)
+- [Design the user experience][Design the user experience]
+- [Add media to messages](https://docs.microsoft.com/en-us/azure/bot-service/bot-builder-howto-add-media-attachments?view=azure-bot-service-4.0&tabs=javascript)
+- [Rich card types](https://docs.microsoft.com/en-us/azure/bot-service/rest-api/bot-framework-rest-connector-add-rich-cards?view=azure-bot-service-4.0)
+
+  [Design the user experience]: https://docs.microsoft.com/en-us/azure/bot-service/bot-service-design-user-experience?view=azure-bot-service-4.0#cards
