+++ samples/javascript_nodejs/06.using-cards/deploymentScripts/msbotClone/bot.recipe
@@ -1,0 +1,11 @@
+{
+  "version": "1.0",
+  "resources": [
+    {
+      "type": "endpoint",
+      "id": "90",
+      "name": "development",
+      "url": "http://localhost:3978/api/messages"
+    }
+  ]
+}
