+++ samples/javascript_nodejs/06.using-cards/using-cards.bot
@@ -1,0 +1,15 @@
+{
+    "name": "using-cards",
+    "services": [
+        {
+            "type": "endpoint",
+            "name": "development",
+            "endpoint": "http://localhost:3978/api/messages",
+            "appId": "",
+            "appPassword": "",
+            "id": "90"
+        }
+    ],
+    "padlock": "",
+    "version": "2.0"
+}
